Build/requirements.txt
adal==1.2.1
applicationinsights==0.11.9
asn1crypto==0.24.0
azure-common==1.1.22
azure-functions==1.0.0b4
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.52.0
azure-mgmt-containerregistry==2.8.0
azure-mgmt-keyvault==1.1.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==2.2.0
azure-mgmt-storage==3.3.0
azure-nspkg==3.0.2
azureml==0.2.7
azureml-contrib-notebook==1.0.43.*
azureml-core==1.0.43
azureml-dataprep==1.1.5
azureml-dataprep-native==13.0.0
azureml-pipeline==1.0.43
azureml-pipeline-core==1.0.43
azureml-pipeline-steps==1.0.43
azureml-sdk==1.0.43
azureml-telemetry==1.0.43
azureml-train==1.0.43
azureml-train-core==1.0.43
azureml-train-restclients-hyperdrive==1.0.43
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.6
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
contextlib2==0.5.5
cryptography==2.7
distro==1.4.0
docker==4.0.1
dotnetcore2==2.1.8
idna==2.8
isodate==0.6.0
jeepney==0.4
jmespath==0.9.4
jsonpickle==1.2
msrest==0.6.6
msrestazure==0.6.1
ndg-httpsclient==0.5.1
numpy==1.16.4
oauthlib==3.0.1
pandas==0.24.2
paramiko==2.5.0
pathspec==0.5.9
pyasn1==0.4.5
pycparser==2.19
PyJWT==1.7.1
PyNaCl==1.3.0
pyOpenSSL==19.0.0
python-dateutil==2.8.0
pytz==2019.1
pytest
requests==2.22.0
requests-oauthlib==1.2.0
ruamel.yaml==0.15.89
SecretStorage==3.1.1
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0